* 'android-bridge-2.6.29' of git://dev.omapzoom.org/pub/scm/tidspbridge/kernel-dspbridge:
  DSPBRIDGE: Fix wrong condition during shared message processing
  DSPBRIDGE: Get dsp max opp using dsp opp table instead of using omap_rev_id
  DSPBRIDGE: Change wrong ifndef condition
  DSPBRIDGE: Bump OPP to minimal needed by DSP before running.
  DSPBRIDGE: fill OPP table in shared memory depending on silicon
  DSPBRIDGE: Avoid REGistry if pDevContext is available

+++ drivers/dsp/bridge/rmgr/drv.c
@@ -1359,7 +1359,7 @@ static DSP_STATUS RequestBridgeResourcesDSP(u32 dwContext, s32 bRequest)
 							OMAP_DSP_MEM3_SIZE);
 		pResources->dwPerBase = ioremap(OMAP_PER_CM_BASE,
 							OMAP_PER_CM_SIZE);
-               pResources->dwPerPmBase = (u32)ioremap(OMAP_PER_PRM_BASE,
+               pResources->dwPerPmBase = ioremap(OMAP_PER_PRM_BASE,
                                                        OMAP_PER_PRM_SIZE);
                pResources->dwCorePmBase = (u32)ioremap(OMAP_CORE_PRM_BASE,
                                                        OMAP_CORE_PRM_SIZE);
